+++ server/commands/read_authzmodels.go
@@ -1,4 +1,4 @@
-package queries
+package commands
 
 import (
 	"context"